deleted duplicate files. renamed 'app' to 'src' to resolve errors with variable names in testing folder
