--- testing/README 2.md
@@ -1,10 +1,0 @@
-This folder is for backend testing. Created by Lucas Videtto of T38
-
-To run a specific test file, execute the following line in a terminal:
-    pytest 'path/to/test/file/eg/login_test.py'
-
-or simply run pytest to execute all tests
-
-ENSURE BACKEND AND FRONTEND SERVER ARE RUNNING TO PERFORM TESTS !!!
-
-PYTEST MUST BE INSTALLED ON LOCAL SYSTEM !!!
